code/requirements.txt
boto3==1.17.71
botocore==1.20.71
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
Flask==2.0.0
idna==2.10
itsdangerous==2.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Jinja2==3.0.0
jmespath==0.10.0
MarkupSafe==2.0.0
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
Werkzeug==2.0.0
python-Levenshtein==0.12.2
